cleanup(lib): move equals Matcher and CustomMatcher into their own libs
Reduces lines in lib/src/core_matches.dart by half

No public API change
