+++ lib/src/iterable_matchers.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'core_matchers.dart';
+import 'equals_matcher.dart';
 import 'description.dart';
 import 'interfaces.dart';
 import 'util.dart';
