+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: matcher
-version: 0.12.2+1
+version: 0.12.3-dev
 author: Dart Team <misc@dartlang.org>
 description: Support for specifying test expectations
 homepage: https://github.com/dart-lang/matcher
